:tada: 3.2.0.RELEASE 新增灵活数据权限特性

+++ src/test/java/org/springblade/test/BladeTest.java
@@ -21,7 +21,7 @@ public class BladeTest {
 
 	@Test
 	public void contextLoads() {
-		int count = noticeService.count();
+		Long count = noticeService.count();
 		System.out.println("notice数量：[" + count + "] 个");
 	}
